requirements.txt
allure-python-commons ==2.15.0
requests==2.32.5 # Библиотека для работы с HTTP-запросами
pytest==8.4.2 # Фреймворк для написания и запуска тестов
pytest-html==4.1.1 # Плагин для генерации HTML-отчетов
pytest-xdist==3.8.0 # Плагин для параллельного выполнения тестов
allure-pytest==2.15.0 # Генерация allure отчета
